Create KHSTREAM
This is the viewcontroller.swift file that is attached to a storyboard.main file. 
username and password field are text fields that should be verified against a database of user accounts
Login button will verify username and password fields. IF Blanks should bring up error message

launchEmail is a button that should auto create an email to the correct address for technical support